+++ crates/zng-app/src/render.rs
@@ -74,6 +74,9 @@ pub trait Img {
     fn alpha_type(&self) -> zng_view_api::AlphaType {
         zng_view_api::AlphaType::PremultipliedAlpha
     }
+
+    /// Image pixel size.
+    fn size(&self) -> PxSize;
 }
 
 macro_rules! expect_inner {
@@ -1275,6 +1278,7 @@ impl FrameBuilder {
         &mut self,
         bounds: PxRect,
         widths: PxSideOffsets,
+        slice: PxSideOffsets,
         fill: bool,
         repeat_horizontal: RepeatMode,
         repeat_vertical: RepeatMode,
@@ -1290,6 +1294,8 @@ impl FrameBuilder {
                 bounds,
                 NinePatchSource::Image { image_id, rendering },
                 widths,
+                image.size(),
+                slice,
                 fill,
                 repeat_horizontal,
                 repeat_vertical,
@@ -1303,6 +1309,7 @@ impl FrameBuilder {
         &mut self,
         bounds: PxRect,
         widths: PxSideOffsets,
+        slice: PxSideOffsets,
         fill: bool,
         repeat_horizontal: RepeatMode,
         repeat_vertical: RepeatMode,
@@ -1329,6 +1336,8 @@ impl FrameBuilder {
                     stops: stops.to_vec().into_boxed_slice(),
                 },
                 widths,
+                bounds.size,
+                slice,
                 fill,
                 repeat_horizontal,
                 repeat_vertical,
@@ -1342,6 +1351,7 @@ impl FrameBuilder {
         &mut self,
         bounds: PxRect,
         widths: PxSideOffsets,
+        slice: PxSideOffsets,
         fill: bool,
         repeat_horizontal: RepeatMode,
         repeat_vertical: RepeatMode,
@@ -1372,6 +1382,8 @@ impl FrameBuilder {
                     stops: stops.to_vec().into_boxed_slice(),
                 },
                 widths,
+                bounds.size,
+                slice,
                 fill,
                 repeat_horizontal,
                 repeat_vertical,
@@ -1385,6 +1397,7 @@ impl FrameBuilder {
         &mut self,
         bounds: PxRect,
         widths: PxSideOffsets,
+        slice: PxSideOffsets,
         fill: bool,
         repeat_horizontal: RepeatMode,
         repeat_vertical: RepeatMode,
@@ -1415,6 +1428,8 @@ impl FrameBuilder {
                     stops: stops.to_vec().into_boxed_slice(),
                 },
                 widths,
+                bounds.size,
+                slice,
                 fill,
                 repeat_horizontal,
                 repeat_vertical,