+++ crates/zng-view-api/src/display_list.rs
@@ -258,11 +258,14 @@ impl DisplayListBuilder {
     }
 
     /// Push a nine-patch border.
+    #[expect(clippy::too_many_arguments)]
     pub fn push_nine_patch_border(
         &mut self,
         bounds: PxRect,
         source: NinePatchSource,
         widths: PxSideOffsets,
+        img_size: PxSize,
+        slice: PxSideOffsets,
         fill: bool,
         repeat_horizontal: RepeatMode,
         repeat_vertical: RepeatMode,
@@ -271,6 +274,8 @@ impl DisplayListBuilder {
             bounds,
             source,
             widths,
+            img_size,
+            slice,
             fill,
             repeat_horizontal,
             repeat_vertical,
@@ -830,6 +835,8 @@ pub enum DisplayItem {
         bounds: PxRect,
         source: NinePatchSource,
         widths: PxSideOffsets,
+        img_size: PxSize,
+        slice: PxSideOffsets,
         fill: bool,
         repeat_horizontal: RepeatMode,
         repeat_vertical: RepeatMode,